Store book covers using ISBN as file name.
Previously, the book covers were stored with a UUID.  But that makes it
tough to avoid duplicating book covers if the library cache is deleted
for whatever reason.  It also makes it difficult to manage the files on
disk (since it's harder to tell which books go with which covers).

This changeset 1) alters the plugin to store the book covers using the
ISBN as the filename, and 2) only downloads the cover if the cover
doesn't exist.

And, finally, while I was at it, I decided to add the publisher and
publication date metadata to the front matter.