Switch to using Calibre as data source.
The previous version of this plugin used the fetch-ebook-metadata tool
to retrieve information about a book.  The trouble with that method is,
unattended, the quality of the metadata that is retrieved is highly
variable.  And to clean it up, you basically have to hack around in the
plugin cache.

This changeset completely alters the way the plugin functions.  Now,
instead of using fetch-ebook-metadata, the plugin interrogates a local
Calibre database to receive ebook metadata and the cover image.

The advantage to this method is you can groom the metadata in Calibre
and then rebuild the site, and it'll use that data.  So you get all the
organizational tools in Calibre at your disposal.

The downside is you have to maintain a Calibre database.

But for me, that's fine, so that's the way this works now.

The plugin now uses this integration to also pull series information, as
well as star ratings, from Calibre.  The latter is particularly useful
because you can rate books in Calibre and then embed that rating in site
pages.

And, as a bonus, we can simplify the logic and kill the on-disk
persistent cache (though we still maintain an in-memory cache for
performance reasons), since Calibre *is* that cache.